NXT-930 Rename TracEmbed to StoryblokTracEmbed
