+++ web/src/components/Storyblok/StoryblokTracEmbed/ClientTracEmbed.tsx
@@ -3,8 +3,7 @@ import dynamic from "next/dynamic";
 const LoadingPlaceholder = () => <div>Loading Trac Jobs...</div>;
 
 export const ClientTracEmbed = dynamic(
-	/* webpackChunkName: "infogram-embed" */
-	() => import("./TracEmbed").then((mod) => mod.TracEmbed),
+	() => import("./StoryblokTracEmbed").then((mod) => mod.StoryblokTracEmbed),
 	{
 		ssr: false,
 		loading: () => <LoadingPlaceholder />,
